+++ packages/nsid/lib/nsid.dart
@@ -4,3 +4,4 @@
 
 export 'package:nsid/src/invalid_nsid_error.dart';
 export 'package:nsid/src/nsid.dart';
+export 'package:nsid/src/validation.dart';