+++ packages/nsid/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.0.2
 
 - Refactored dartdoc.
+- Exposed `ensureValidNsid`. ([#140](https://github.com/myConsciousness/atproto.dart/issues/140))
 
 ## v0.0.1